Project/requirements.txt
amqp==5.3.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
async-timeout==5.0.1
billiard==4.2.1
celery==5.4.0
cffi==1.17.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.6.10
cron-descriptor==1.4.5
cryptography==44.0.1
Django==4.2.17
django-cron==0.6.0
django-encrypted-model-fields==0.6.5
django-environ==0.11.2
django-timezone-field==7.1
environ==1.0
exceptiongroup==1.2.2
factory_boy==3.3.1
Faker==33.3.1
iniconfig==2.0.0
kombu==5.4.2
nulltype==2.3.1
packaging==24.2
plaid-python==28.0.0
pluggy==1.5.0
prompt_toolkit==3.0.50
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycparser==2.22
pytest==8.3.4
pytest-django==4.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
redis==5.2.1
six==1.17.0
sqlparse==0.5.2
tomli==2.2.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
